refactor: let appearance own its wiring
Startup had to know which modules inside appearance register QML
types, so every type added or moved there meant touching startup too.
Now the feature exposes one entry point and the root just calls it.

Same for the bindings, which the root can now reach without digging
into the feature. While in there, the startup hook got the name that
says what it actually does.
