+++ mpvqc/injections.py
@@ -5,7 +5,7 @@
 import inject
 
 import mpvqc.services as s
-from mpvqc.appearance.injections import bindings as appearance_bindings
+from mpvqc.appearance import bindings as appearance_bindings
 
 
 def _settings_service() -> s.SettingsService:
