+++ pyproject.toml
@@ -128,7 +128,6 @@ files = [
     "mpvqc/__init__.py",
     "mpvqc/appearance/__init__.py",
     "mpvqc/appearance/domain.py",
-    "mpvqc/appearance/injections.py",
     "mpvqc/appearance/models/__init__.py",
     "mpvqc/appearance/models/accent_colors.py",
     "mpvqc/appearance/models/color_scheme_preference.py",
@@ -139,6 +138,7 @@ files = [
     "mpvqc/appearance/viewmodels/__init__.py",
     "mpvqc/appearance/viewmodels/dialog.py",
     "mpvqc/appearance/viewmodels/palette.py",
+    "mpvqc/appearance/wiring.py",
     "mpvqc/application.py",
     "mpvqc/build.py",
     "mpvqc/close_event_filter.py",
